fixing bugs

+++ src/main/java/com/example/moneyapp2/web/IncomeController.java
@@ -1,11 +1,13 @@
 package com.example.moneyapp2.web;
 
 import com.example.moneyapp2.model.dto.income.CreateIncomeDTO;
+import com.example.moneyapp2.model.dto.income.EditIncomeDTO;
 import com.example.moneyapp2.model.dto.income.IncomeDetailsDTO;
 import com.example.moneyapp2.model.dto.income.IncomeInfoDTO;
 import com.example.moneyapp2.service.IncomeService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,31 +33,59 @@ public class IncomeController {
     }
 
     @GetMapping("/details/{id}")
-    public ResponseEntity<IncomeDetailsDTO> detailedIncomeInfo(@PathVariable Long id) {
+    public ResponseEntity<IncomeDetailsDTO> detailedIncomeInfo(@PathVariable Long id, Principal principal) {
 
         if(this.incomeService.IncomeNotPresent(id)) {
             return ResponseEntity.notFound().build();
         }
 
+        if(this.incomeService.unauthorizedUser(id, principal.getName())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(this.incomeService.getDetailsOfIncome(id));
+    }
+
+    @GetMapping("/edit/{id}")
+    public ResponseEntity<EditIncomeDTO> editIncomeInfo(@PathVariable Long id,
+                                                        Principal principal) {
+
+        if(this.incomeService.IncomeNotPresent(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if(this.incomeService.unauthorizedUser(id, principal.getName())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        return ResponseEntity.ok(this.incomeService.getSingleIncome(id));
     }
 
     @PatchMapping("/edit/{id}")
     public ResponseEntity<IncomeDetailsDTO> editIncomeInfo(@PathVariable Long id,
-                                                          @Valid @RequestBody CreateIncomeDTO changedIncomeInfo) {
+                                                          @Valid @RequestBody CreateIncomeDTO changedIncomeInfo,
+                                                           Principal principal) {
 
         if(this.incomeService.IncomeNotPresent(id)) {
             return ResponseEntity.notFound().build();
+        }
+
+        if(this.incomeService.unauthorizedUser(id, principal.getName())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
 
         return ResponseEntity.ok(this.incomeService.editIncome(id, changedIncomeInfo));
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deleteIncome(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteIncome(@PathVariable Long id, Principal principal) {
 
         if(this.incomeService.IncomeNotPresent(id)) {
             return ResponseEntity.notFound().build();
+        }
+
+        if(this.incomeService.unauthorizedUser(id, principal.getName())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
 
         this.incomeService.deleteIncome(id);
